Load SSH identities on-the-fly
OS X Sierra no longer loads SSH keys by default at login, so do it the first shell we open. Running `ssh-add -A` adds about half a second to the shell boot time, so wrap it in a check to see if we need to run it. (`ssh-add -l` time difference is negligible by comparison.)