Add a git commit template
Writing git commits is easy to do badly, and not always explain yourself.

In an effort to train myself into writing better commit messages, lets add this commit template to encourage me to follow it. Taken from <https://netty.io/wiki/writing-a-commit-message.html>

This will cause git commit to pop open with the template pre-filled in as a prompt.

The starting newline is required to get TextMate to syntax highlight the diff correctly.